src/module/users/useCases/deleteUserUseCase/index.ts
import { UserNotFoundError } from '@module/users/errors/UserNotFoundError'
import { UsersRepository } from '@module/users/repositories/contracts/UsersRepository'
import { Either, left, right } from '@shared/core/errors/Either'

interface Request {
  id: string
}

type Response = Either<UserNotFoundError, null>

export class DeleteUserUseCase {
  constructor(private readonly usersRepository: UsersRepository) {}

  async execute({ id }: Request): Promise<Response> {
    const user = await this.usersRepository.findById(id)

    if (!user) {
      return left(new UserNotFoundError())
    }

    await this.usersRepository.delete(id)
    return right(null)
  }
}
